Moved `Constants` to core
