+++ app/src/main/java/com/imashnake/animite/features/home/HomeViewModel.kt
@@ -10,7 +10,7 @@ import com.imashnake.animite.core.data.Resource
 import com.imashnake.animite.core.data.Resource.Companion.asResource
 import com.imashnake.animite.dev.ext.nextSeason
 import com.imashnake.animite.dev.ext.season
-import com.imashnake.animite.dev.internal.Constants
+import com.imashnake.animite.core.Constants
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.SharingStarted
