+++ app/src/main/java/com/imashnake/animite/features/ui/MediaSmall.kt
@@ -35,7 +35,7 @@ import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.imashnake.animite.api.anilist.MediaListQuery
 import com.imashnake.animite.api.anilist.type.MediaType
-import com.imashnake.animite.dev.internal.Constants.CROSSFADE_DURATION
+import com.imashnake.animite.core.Constants.CROSSFADE_DURATION
 import com.imashnake.animite.R
 import com.imashnake.animite.core.ui.LocalPaddings
