+++ core/src/main/kotlin/com/imashnake/animite/core/Constants.kt
@@ -1,4 +1,4 @@
-package com.imashnake.animite.dev.internal
+package com.imashnake.animite.core
 
 object Constants {
     const val ANILIST_BASE_URL = "https://graphql.anilist.co/"
